Cocoa Port: - New feature: If the user quits DeSmuME while display windows remain, the user has the option to restore the display windows on app launch.

+++ desmume/src/cocoa/userinterface/DisplayWindowController.h
@@ -70,7 +70,7 @@
 	GLubyte vtxIndexBuffer[12];
 	
 	BOOL isShaderSupported;
-	unsigned int vtxBufferOffset;
+	size_t vtxBufferOffset;
 }
 
 @property (retain) InputManager *inputManager;
